+++ stanzas/treemap/shadeColor.js
@@ -14,9 +14,12 @@ export default function (color, percent) {
   G = G < 255 ? G : 255;
   B = B < 255 ? B : 255;
 
-  let RR = R.toString(16).length === 1 ? "0" + R.toString(16) : R.toString(16);
-  let GG = G.toString(16).length === 1 ? "0" + G.toString(16) : G.toString(16);
-  let BB = B.toString(16).length === 1 ? "0" + B.toString(16) : B.toString(16);
+  const RR =
+    R.toString(16).length === 1 ? "0" + R.toString(16) : R.toString(16);
+  const GG =
+    G.toString(16).length === 1 ? "0" + G.toString(16) : G.toString(16);
+  const BB =
+    B.toString(16).length === 1 ? "0" + B.toString(16) : B.toString(16);
 
   return "#" + RR + GG + BB;
 }
